refactor(gui): move main window class to its own file
